Remove private registry publication script
This module was originally published to my personal Terraform Cloud
private module registry using the registry API and a custom script.
This module is now published to the public registry here:
https://registry.terraform.io/modules/br3ndonland/github-actions-oidc/aws/latest

As the module is published to the public registry, the script and its
corresponding GitHub Actions workflow job are not needed.
This commit will remove them.

The script is still available in a public GitHub Gist here:
https://gist.github.com/br3ndonland/2e3665a8117b5594b00f3d556d33cd57