Move minimum required IAM permissions to README
This repo previously had an `examples/iam` directory containing minimum
required permissions for applying configurations using the module. The
`examples/` directory is usually for Terraform configuration examples,
not auxiliary IAM policies, so the IAM example did not appear in
Terraform registry docs.

This commit will move the minimum required permissions to the README and
remove the `examples/iam` directory.