Clarify module publication status
The GitHub repo was temporarily renamed to tofu-aws-github-actions-oidc,
but had to be changed back to terraform-aws-github-actions-oidc in order
to claim the HCP Terraform public registry namespace. The README has
been updated to clarify that the module will continue to be published to
the HCP Terraform public registry. Apologies for any confusion this may
have caused.

https://developer.hashicorp.com/terraform/cloud-docs/public-namespace